Extract Range type from ModeSubsidy; extract test into a Spec
